% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchcase.r
\name{switchcase}
\alias{switchcase}
\title{Package 'switchcase'}
\description{
A simple yet flexible switch-case construct for R. Function
  \code{\link{switchCase}()} provides the actual switch-case functionality.
  The \code{\link{alt}()} function is used to formulate alternative 'case'
  branches of the switch-case construct.
}
